Major redesign. Web Scraping instead of API calls
Completely redesigned the methods of collecting data. Instead of using trial subscription to make API calls, this update is scraping the data directly from Gas Buddy. Also, the results are more specific to the median gas price of a zip code, instead of an average value for an entire city.